--- a/Cargo.toml
+++ b/Cargo.toml
@@ -4,11 +4,6 @@ members = [
   "apps/sequencer",
   "apps/sequencer_tests",
   "apps/trigger-oracle",
-  "apps/oracles/cex-price-feeds",
-  "apps/oracles/exsat-holdings",
-  "apps/oracles/gecko-terminal",
-  "apps/oracles/eth-rpc",
-  "apps/oracles/stock-price-feeds",
   "libs/alpaca_core",
   "libs/anomaly_detection",
   "libs/data_feeds",
@@ -51,7 +46,7 @@ blocksense-metrics = { path = "libs/metrics" }
 anyhow = "1"
 futures = "0.3"
 serde = { version = "1.0", features = ["derive"] }
-serde_json = { version = "1.0", features = ["float_roundtrip"]}
+serde_json = { version = "1.0", features = ["float_roundtrip"] }
 url = "2.5"
 log = "0.4"
 rand = "0.8"