--- a/apps/oracles/.cargo/config.toml
+++ b/apps/oracles/.cargo/config.toml
@@ -1,0 +1,2 @@
+[build]
+target = "wasm32-wasip1"