--- a/apps/oracles/gecko-terminal/spin.toml
+++ b/apps/oracles/gecko-terminal/spin.toml
@@ -65,7 +65,7 @@ decimals = 8
 data = '{"pair":{"base":"oLP","quote":"USD"},"decimals":8,"category":"Crypto","market_hours":"Crypto","arguments":[{"network":"arbitrum","pool":"0x09e6a7adfb8c06f22adfefd872de73059d955be3","reverse":false}]}'
 
 [component.gecko-terminal]
-source = "../../../target/wasm32-wasip1/release/gecko_terminal.wasm"
+source = "../target/wasm32-wasip1/release/gecko_terminal.wasm"
 allowed_outbound_hosts = ["https://api.geckoterminal.com"]
 
 [component.gecko-terminal.build]