--- a/apps/oracles/stock-price-feeds/spin.toml
+++ b/apps/oracles/stock-price-feeds/spin.toml
@@ -29,7 +29,7 @@ decimals = 8
 data = '{"pair":{"base":"SPY","quote":"USD"},"decimals":8,"category":"Equities","market_hours":"US_Equities","arguments":{"providers":["AlpacaMarkets","AlphaVantage","YahooFinance","twelvedata","FMP"]}}'
 
 [component.stock-price-feeds]
-source = "../../../target/wasm32-wasip1/release/stock_price_feeds.wasm"
+source = "../target/wasm32-wasip1/release/stock_price_feeds.wasm"
 allowed_outbound_hosts = [
   "https://data.alpaca.markets",
   "https://www.alphavantage.co",