--- a/nix/pkgs/oracle-script/default.nix
+++ b/nix/pkgs/oracle-script/default.nix
@@ -1,35 +1,27 @@
 {
-  lib,
   craneLib,
   pkg-config,
   filesets,
   version ? "dev",
-  oracle-path,
-
-  # whether the given oracle script is part of the main workspace
-  # or it's standalone
-  standalone,
 }:
 let
-  oracle-name = builtins.baseNameOf oracle-path;
 
   sharedAttrs = rec {
-    pname = "oracle-script-${oracle-name}";
+    pname = "blocksense-oracle-scripts";
     inherit (filesets.rustSrc) src;
-
-    postUnpack = lib.optionalString standalone ''
-      cd $sourceRoot/${builtins.toString oracle-path}
-      sourceRoot="."
-    '';
-
-    cargoToml = ../../.. + oracle-path + /Cargo.toml;
 
     nativeBuildInputs = [
       pkg-config
     ];
 
-    cargoExtraArgs =
-      "--target wasm32-wasip1" + lib.optionalString (!standalone) " --package=${oracle-name}";
+    postUnpack = ''
+      cd $sourceRoot/apps/oracles
+      sourceRoot="."
+    '';
+
+    cargoToml = ../../../apps/oracles/Cargo.toml;
+    cargoLock = ../../../apps/oracles/Cargo.lock;
+    cargoExtraArgs = "--target wasm32-wasip1";
     doCheck = false;
     strictDeps = true;
   };