--- a/nix/shells/pkg-sets/all.nix
+++ b/nix/shells/pkg-sets/all.nix
@@ -1,6 +1,4 @@
 {
-  self,
-  config,
   ...
 }:
 let
@@ -12,18 +10,9 @@ in
     ./rust.nix
     ./anvil.nix
     ./kafka.nix
-
-    self.nixosModules.blocksense-process-compose
-    ../../test-environments/example-setup-01.nix
   ];
 
   enterShell = ''
     git clean -fdx -- ${generated-cfg-dir}
-
-    ln -fs ${config.process.managers.process-compose.configFile} "${generated-cfg-dir}/process-compose.yml"
-
-    for file in "${config.services.blocksense.config-dir}"/*; do
-      ln -s "$file" "${generated-cfg-dir}/$(basename "$file")"
-    done
   '';
 }