--- a/nix/shells/pkg-sets/dev-shell.nix
+++ b/nix/shells/pkg-sets/dev-shell.nix
@@ -15,6 +15,7 @@
     oras
     just
     findutils
+    process-compose
   ];
 
   enterShell =