## Optimize Blocksense Environment Bootstrapping

This PR makes several improvements to the developer experience and performance of the `blocksense` process-compose environments:

---

### Key Changes

#### 1. Replace `start-blocksense` with `start-environment`

* Introduces a unified `start-environment` command.
* It builds both the Blocksense binaries and the environment, then launches `process-compose` with the selected environment.
* This replaces the fragmented `start-blocksense` and reduces ambiguity in workflows.
* Usage: `just start-environment <env>`

#### 2. Speed Up Oracle Build Time

* The previous `mkOracleScript` function was a significant bottleneck (5–6s per oracle).
* Optimization steps:

  * Oracles are now defined as a separate workspace.
  * All oracle scripts are built in a single derivation, massively reducing build overhead.

#### 3. Improved Environment Configuration Structure

* Environment artifacts live in a separate directory under `config/generated/process-compose`
* Each environment folder now includes its own `sequencer` and `reporter` config files.
* This promotes clarity and improves encapsulation of environment-specific logic.
* For example, if you build all environments with `just build-environment`, this will produce:
<img src="https://github.com/user-attachments/assets/15937b74-5e79-4343-a7c6-ec99567a72e8" width="300" />


#### 4. `process-compose` Package and Shell Changes

* `process-compose` is now explicitly included in the `devShell`.
* Removed implicit dependency on it from the default shell.

---

### Impact

* Local environment startup is now significantly faster.
* Starting blocksense with `process-compose` now just requires running `just start-environment <env>`

### Time Improvements Results

> All times are measured on my laptop. I expect the times on the office PCs to be better

#### `direnv reload` when using `default` shell
> Time measure with `time nix print-dev-env --impure .#devShells.x86_64-linux.default `

| **Before** | **After** |
|-----------|-----------|
|    ![image](https://github.com/user-attachments/assets/b5085915-25fd-432c-9718-0194f2d0fec5)  |    ![image](https://github.com/user-attachments/assets/da6f4cb4-8855-4dec-86f2-d33ab5f7702f) |
|     ~ 47.35 secs       |     ~  7.27 secs      |


#### `just build-environment <env>` ( with the assumption that the rust code has been built before )
> Time measure with   `time just build-environment example-setup-01 `

| **Before** | **After** |
|-----------|-----------|
|    ![image](https://github.com/user-attachments/assets/f8703987-a763-4426-88db-c705099be9ba)|    ![image](https://github.com/user-attachments/assets/20b26167-24be-4f97-83c7-2933c95a70dc) |
|     ~ 39.32 secs       |     ~  7.70 secs      |

#### Start Blocksense environment on a clean setup

> Before tests, I deleted all artifacts with `just clean`

| **Before** | **After** |
|-----------|-----------|
|       `direnv reload`  +  `just build-blocksense`  +`just build-environment example-setup-02` + `just start-environment example-setup-02`    |    `direnv reload` + `just start-environment example-setup-02` ( all required building are dependencies of that command )  |
|     ~ 278.16 secs = 47.35 +  198 + 32.8 + 0.01  |     ~  140 secs = 7.70 + 132     |

### Next steps
- [ ] Update repo documentation [BSN-3186: Update Running the System section in SETUP.md ](https://coda.io/d/_d6vM0kjfQP6#_tuk5j/r3186&view=full)
- [ ] Get rid of `--impure`

**Big thanks to @reo101 & @PetarKirov for the time spent brainstorming `nix` with me!**